move version 1.x message case classes to a messages.v1x package
