+++ ocpp-j-api/src/test/scala/com/thenewmotion/ocpp/json/api/RequestHandlerSpec.scala
@@ -5,7 +5,8 @@ package api
 import scala.concurrent.duration.DurationInt
 import scala.concurrent.ExecutionContext.Implicits.global
 import org.specs2.mutable.Specification
-import com.thenewmotion.ocpp.messages._
+import com.thenewmotion.ocpp.messages.v1x._
+import ReqRes._
 import RequestHandler._
 
 import scala.concurrent.Await
