+++ ocpp-json/src/main/scala/com/thenewmotion/ocpp/json/OcppJ.scala
@@ -3,7 +3,7 @@ package json
 
 import org.json4s._
 import native.JsonMethods.{compact, render}
-import messages.Message
+import messages.v1x.Message
 
 
 /** Reading and writing OCPP 1.x messages encoded with JSON */
@@ -14,7 +14,7 @@ object OcppJ {
   def serialize[M <: Message, V <: Version](msg: M)(implicit versionVariant: OcppMessageSerializer[M, V]): JValue =
     versionVariant.serialize(msg)
 
-  def write[M <: messages.Message, V <: Version](msg: M)(implicit versionVariant: OcppMessageSerializer[M, V]): String =
+  def write[M <: Message, V <: Version](msg: M)(implicit versionVariant: OcppMessageSerializer[M, V]): String =
     compact(render(serialize[M, V](msg)))
 
   def deserialize[M <: Message, V <: Version](json: JValue)(implicit versionVariant: OcppMessageSerializer[M, V]): M =
