--- ocpp-messages/src/main/scala/com/thenewmotion/ocpp/messages/package.scala
@@ -1,5 +1,0 @@
-package com.thenewmotion.ocpp
-
-package object messages {
-  type IdTag = String
-}
