+++ ocpp-messages/src/main/scala/com/thenewmotion/ocpp/messages/v1x/CentralSystemAction.scala
@@ -1,4 +1,4 @@
-package com.thenewmotion.ocpp.messages
+package com.thenewmotion.ocpp.messages.v1x
 
 import enums.reflection.EnumUtils.{Enumerable, Nameable}
