+++ ocpp-messages/src/main/scala/com/thenewmotion/ocpp/messages/v1x/DataTransferStatus.scala
@@ -1,5 +1,4 @@
-package com.thenewmotion.ocpp
-package messages
+package com.thenewmotion.ocpp.messages.v1x
 
 import enums.reflection.EnumUtils.{Enumerable, Nameable}
