+++ ocpp-messages/src/main/scala/com/thenewmotion/ocpp/messages/v1x/Scope.scala
@@ -1,5 +1,4 @@
-package com.thenewmotion.ocpp
-package messages
+package com.thenewmotion.ocpp.messages.v1x
 
 /**
  * @author Yaroslav Klymko
