+++ ocpp-messages/src/main/scala/com/thenewmotion/ocpp/messages/v1x/SyncCentralSystem.scala
@@ -1,4 +1,6 @@
-package com.thenewmotion.ocpp.messages
+package com.thenewmotion.ocpp
+package messages
+package v1x
 
 trait SyncCentralSystem {
   def authorize(req: AuthorizeReq): AuthorizeRes
