+++ ocpp-messages/src/main/scala/com/thenewmotion/ocpp/messages/v1x/v1x.scala
@@ -1,0 +1,5 @@
+package com.thenewmotion.ocpp.messages
+
+package object v1x {
+  type IdTag = String
+}
